add touch support for drag&drop and split components
