+++ src/components/settings/SettingsDashboardTabDesktop.vue
@@ -1,0 +1,110 @@
+<style scoped>
+    .ghost {
+        opacity: 0.5;
+        background: #c8ebfb;
+    }
+</style>
+
+<template>
+    <v-card flat>
+        <v-card-text>
+            <v-row>
+                <v-col class="col-12 col-md-4">
+                    <v-card class="mx-auto" max-width="300" tile >
+                        <v-list dense>
+                            <v-list-item>
+                                <v-list-item-icon>
+                                    <v-icon>mdi-information</v-icon>
+                                </v-list-item-icon>
+                                <v-list-item-content>
+                                    <v-list-item-title>{{ $t('Panels.StatusPanel.Headline') }}</v-list-item-title>
+                                </v-list-item-content>
+                            </v-list-item>
+                            <draggable v-model="desktopLayout1" class="v-list-item-group" ghost-class="ghost" group="desktopViewport">
+                                <template v-for="(element) in desktopLayout1">
+                                    <v-list-item :key="'item-desktop-'+element.name" link>
+                                        <v-list-item-icon>
+                                            <v-icon v-text="convertToIcon(element.name)"></v-icon>
+                                        </v-list-item-icon>
+                                        <v-list-item-content>
+                                            <v-list-item-title>{{ $t('Panels.'+capitalize(element.name)+'Panel.Headline') }}</v-list-item-title>
+                                        </v-list-item-content>
+                                    </v-list-item>
+                                </template>
+                            </draggable>
+                        </v-list>
+                    </v-card>
+                </v-col>
+                <v-col class="col-12 col-md-4">
+                    <v-card class="mx-auto" max-width="300" tile >
+                        <v-list dense>
+                            <draggable v-model="desktopLayout2" class="v-list-item-group" ghost-class="ghost" group="desktopViewport">
+                                <template v-for="(element) in desktopLayout2">
+                                    <v-list-item :key="'item-desktop-'+element.name" link>
+                                        <v-list-item-icon>
+                                            <v-icon v-text="convertToIcon(element.name)"></v-icon>
+                                        </v-list-item-icon>
+                                        <v-list-item-content>
+                                            <v-list-item-title>{{ $t('Panels.'+capitalize(element.name)+'Panel.Headline') }}</v-list-item-title>
+                                        </v-list-item-content>
+                                    </v-list-item>
+                                </template>
+                            </draggable>
+                        </v-list>
+                    </v-card>
+                </v-col>
+            </v-row>
+        </v-card-text>
+    </v-card>
+</template>
+
+<script lang="ts">
+import Component from 'vue-class-component'
+import { Mixins } from 'vue-property-decorator'
+import BaseMixin from '@/components/mixins/base'
+import draggable from 'vuedraggable'
+import {capitalize} from "@/plugins/helpers";
+@Component( {
+    components: {
+        draggable
+    }
+}
+)
+export default class SettingsDashboardTabDesktop extends Mixins(BaseMixin) {
+    capitalize = capitalize
+
+    get desktopLayout1() {
+        return this.$store.state.gui?.dashboard?.desktopLayout1?.filter((element: any) => element !== null) ?? []
+    }
+
+    set desktopLayout1(newVal) {
+        newVal = newVal.filter((element: any) => element !== undefined)
+
+        this.$store.dispatch('gui/saveSetting', {name: 'dashboard.desktopLayout1', value: newVal })
+    }
+
+    get desktopLayout2() {
+        return this.$store.state.gui?.dashboard?.desktopLayout2?.filter((element: any) => element !== null) ?? []
+    }
+
+    set desktopLayout2(newVal) {
+        newVal = newVal.filter((element: any) => element !== undefined)
+
+        this.$store.dispatch('gui/saveSetting', {name: 'dashboard.desktopLayout2', value: newVal })
+    }
+
+    convertToIcon(name: string) {
+        switch (name) {
+            case 'webcam': return 'mdi-webcam'
+            case 'zoffset': return 'mdi-arrow-collapse-vertical'
+            case 'control': return 'mdi-gamepad'
+            case 'printsettings': return 'mdi-printer-3d'
+            case 'miscellaneous': return 'mdi-dip-switch'
+            case 'tools': return 'mdi-thermometer-lines'
+            case 'miniconsole': return 'mdi-console-line'
+
+            default: return 'mdi-information'
+        }
+    }
+}
+</script>
